feat(shopping): read brands related to category
After a customer chooses a category for shopping, show only the relevant
brands that exist inside that category. The input is a categoryID, and
the output is a list of related brands by ID and Name. The product
models serve as the basis for finding relevant categories and brands.

Example: When a customer needs to find pens, she will choose
"stationery", then choose "Bic", and finally see a list of models
offered by the Bic company. This can only be achieved by inspecting the
available product models, first by category, and seocnd by brand.