feat(favorite products): Upload CSV file
Upload a CSV file containing two columns. The first column is a thing,
and the second column is a white-space delimited list of details
describing the thing. Return a JSON payload of manipulated content, in
which the white-space is replaced by a comma. The customer can view the
results. The results will be used in a review and confirmation process.